Create simple Overture feature with ID and type
This can be converted into a JSON Schema using the following `linkml`
tool design:

```
$ gen-json-schema schema.yaml
```

I looked for a way to annotate the LinkML so that it would hive the
"type" property off under a top-level "properties" block, as with
GeoJSON, but I don't think LinkML has this capability.

We probably can create an Overture-themed "GeoJSON JSON Schema"
generator that would handle this automatically?