Use higher precision `decimal` instead of `float` for percentage
This is a nitty change and maybe isn't required. I'm partially making
it to ensure we are aware that `decimal` is available as a type, and
partially making it because our percentage-based LR has the downside
of the precision varying based on the length of the segment, so it
theoretically make sense to maintain a higher-precision number.