Model perspectives in names
This commit applies the changes Jonah recently merged into the Overture
schema in PR #319.

https://github.com/OvertureMaps/schema/pull/319

As modeled in this commit, all the perspectives models should be
directly reusable for divisions when I get around to modeling that.

(Which reminds me that one issue with Jonah's PR is that we reviewers
totally failed to push for type reusability so there's duplication.)

One revelation I had while modeling is that the way we do perspectives
is essentially a form of scoping property. That's very interesting.